ttn-kafka-converter/requirements.txt
confluent-kafka
paho-mqtt
cbor2
bitstring
